requirements.txt
blinker==1.7.0
click==8.1.7
dominate==2.9.0
Flask==3.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.3
Pillow==10.1.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
qrcode==7.4.2
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
visitor==0.1.3
Werkzeug==3.0.1
WTForms==3.1.1
